Move operator-related provider items to operators.go
